Add psuedo code
The beginning piece of code to check if certain cells in an excel
document contain any values. This piece is mostly the ground work,
written in psuedo code to outline the bulk of whats to come. Some
initial pieces have been written.

To note noting is working yet besides
a bit of the test code that has been commented out at the very end.