Add elements to place cell values into the arrays
In the 'Do Until' loop, added pieces that push found cell values into
the empty arrays at the specified index. The index is specified to match
the UID columns with the all other values in a given row, even if blank.
This will also end with all array being the same size.
An addition to the CountIndex was specified after which will designate
the next index to be used when a value in found.

Also added in and addition to i to specify to move to the next row.

Small changes: Converted the 'x' into an 'i' to for better loop variable
view.