Fix adding cell value to array
Upon further inspection the an error was being thrown after attempting
to add the value into the array. This has been fixed with 'ReDim
Preserve' before adding values into the specified index of the array.

Note: All arrays in the loop were 'UpComplete' this has been corrected
to specify all the created array names and not iterating the same over
again.