Create excel worksheet
A worksheet in now initialized and created with headers for the given
data that will be displayed. This will be used to map each array into
a given column into order to generate a report.

Note: Currently saving the worksheet is commented out. This was tested,
then removed from being live until after mapping out the arrays is
complete.