Map arrays to worksheet
1. Added more descriptive headers to the columns.
2. Created a way to loop each array and run down a designated column,
moving to the next column when the array is complete. Unfortunately when
tested it bugs out on the first array, after looping through the full
length of the array.

Note: This is strictly a backup entry and is not working code. This will
be picked back up tomorrow and fixed to correctly display the data.