Fix adding array to Excel sheet
Instead of placing a a loop within a loop, used 'Offset'. This will
eliminate the need for the 'x' variable used to set the row on the
Excel sheet. It will now offset down the column filling it with the
defined array, in this case the UID. Not the first for loop was also
updated to a 'Do Until' loop.